move method controller to service and build multithreading process by virtual thread
