+++ src/main/resources/application.yml
@@ -14,6 +14,9 @@ spring:
     show-sql: true
     hibernate:
       ddl-auto: update
+  threads:
+    virtual:
+      enabled: true
 
 # 로깅 설정
 logging:
